refactor: simplify toggle enter
