+++ assets/levels/entry_point.json
@@ -21,7 +21,7 @@
     "test_teleporter_is_teleporting": {
       "data": {
         "type": "level_teleporter",
-        "destination": "test_teleporter_is_teleporting_all_after_activation"
+        "destination": "test_pressure_plate_on_off"
       },
       "position": {
         "x": 0,
