+++ src/main.rs
@@ -53,6 +53,7 @@ struct LevelAssets {
         paths(
             "levels/test_teleporter_is_teleporting.json",
             "levels/test_teleporter_is_teleporting_all_after_activation.json",
+            "levels/test_pressure_plate_on_off.json",
             "levels/entry_point.json",
             "levels/example.json",
             "levels/example_2.json"
